chore(sandbox): restructure
