+++ apps/sandbox/api/index.ts
@@ -1,0 +1,4 @@
+import { handle } from "@hono/node-server/vercel";
+import { app } from "../src";
+
+export default handle(app);
