+++ apps/sandbox/src/index.tsx
@@ -1,9 +1,8 @@
-import { handle } from "@hono/node-server/vercel";
 import { Hono } from "hono";
 import { logger } from "hono/logger";
-import { Page } from "./Page.js";
+import { Page } from "./Page";
 
-const app = new Hono();
+export const app = new Hono();
 
 app.use(logger());
 
@@ -12,5 +11,3 @@ app.get("/", (c) => {
     <Page info={{ limit: 5, used: 1, remaining: 4, resetTime: new Date() }} />,
   );
 });
-
-export default handle(app);
